enforce command-plane architecture gates
